pyproject.toml
[project]
name = "play-pyspark-ai"
version = "0.1.0"
description = "Add a short description here"
authors = [
    { name = "shunta-anezaki", email = "shunta.anezaki@magia.co.jp" }
]
dependencies = [
    "pyspark-ai>=0.1.8",
    "python-dotenv>=1.0.0",
    "pyspark>=3.4.1",
    "pyarrow>=12.0.1",
]
readme = "README.md"
requires-python = ">= 3.8.1"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "ruff>=0.0.276",
]

[tool.hatch.metadata]
allow-direct-references = true
